backend/requirements.txt
fastapi==0.117.1
uvicorn==0.37.0
neomodel==5.5.2
neo4j==5.28.2
python-dotenv==1.1.1
passlib==1.7.4
PyJWT==2.10.1
bcrypt==4.3.0
python-multipart==0.0.18
python-socketio==5.11.4
aiohttp==3.12.15
